Species assessment summary

Get an assessment summary for a particular species (i.e., Latin binomial) or subspecies/variety/subpopulation (i.e., Latin trinomial).

Usage

rl_species(
  genus,
  species,
  infra = NULL,
  subpopulation = NULL,
  key = NULL,
  parse = TRUE,
  ...
)

rl_species_(
  genus,
  species,
  infra = NULL,
  subpopulation = NULL,
  key = NULL,
  ...
)

Arguments

genus

(character) The genus name of the species to look up.

species

(character) The species epithet of the species to look up.

infra

(character) An optional name of the subspecies or variety to look up.

subpopulation

(character) An optional name of the geographically separate subpopulation to look up.

key

(character) An IUCN API token. See rl_use_iucn.

parse

(logical) Whether to parse the output to list (FALSE) or, where possible, data.frame (TRUE). Default: TRUE

...

Curl options passed to the GET request via HttpClient.

Value

A list unless using a function with a trailing underscore, in which case json as character string is returned.

Details

Geographically separate subpopulations of a species are defined as those populations that are so isolated from others of the same species that it is considered extremely unlikely that there is any genetic interchange. In general, listings of such subpopulations are restricted to those that have been isolated for a long period of time.

Assessments of subspecies, varieties, and geographically separate subpopulations must adhere to the same standards as for species assessments. However, these assessments are only included provided there is a global assessment of the species as a whole.

Infraspecific ranks such as formas, subvarieties, cultivars, etc are not included in the Red List.

References

API docs at https://api.iucnredlist.org/.

See also

Examples

if (FALSE) { # \dontrun{
# Get assessment summary for species
ex1 <- rl_species(genus = "Fratercula", species = "arctica")
nrow(ex1$assessments)

# Get assessment summary for subspecies
ex2 <- rl_species(genus = "Gorilla", species = "gorilla",
                  infra = "gorilla")
nrow(ex2$assessments)
} # }