Threat assessment summary

Return a list of the latest assessments which are subject to a specific threat (e.g., energy production and mining, climate change, and severe weather). This will only return assessments for the threat code specified. You will need to do additional requests for sub-threats (e.g., a request for threat code 2_1 will need additional requests for codes 2_1_1, 2_1_2, etc.).

Usage

rl_threats(
  code = NULL,
  key = NULL,
  parse = TRUE,
  all = TRUE,
  page = 1,
  quiet = FALSE,
  ...
)

rl_threats_(code = NULL, key = NULL, all = TRUE, page = 1, quiet = FALSE, ...)

Arguments

code

(character) The code of the threat to look up. If not supplied, a list of all threats will be returned.

key

(character) An IUCN API token. See rl_use_iucn.

parse

(logical) Whether to parse the output to list (FALSE) or, where possible, data.frame (TRUE). Default: TRUE

all

(logical) Whether to retrieve all results at once or not. If TRUE we do the paging internally for you and bind all of the results together. If FALSE, only a single page of results will be retrieved.

page

(integer/numeric) Page to get if all is FALSE. Default: 1. Each page returns up to 100 records. Paging is required because it's too much burden on a server to just "get all the data" in one request.

quiet

(logical) Whether to suppress progress for multi-page downloads or not. Default: FALSE (that is, give progress). Ignored if all = FALSE.

...

Includes the following filters that are supported by the API endpoint (refer to the API docs for more information):

year_published: (integer) Set this to return only assessments from a given year.

latest: (logical) Set this to TRUE to return only the latest assessment for each taxon.

scope_code: (integer) Set this to return only assessments from a particular scope (e.g., 1 for Global, 2 for Europe). This is similar to the region argument of the old Red List API and old versions of rredlist.

Also supports any curl options passed to the GET request via HttpClient.

Value

A list unless using a function with a trailing underscore, in which case json as character string is returned.

References

API docs at https://api.iucnredlist.org/.

See also

Examples

if (FALSE) { # \dontrun{
# Get all threats
rl_threats()
# Get assessment summary for intentional hunting and trapping
rl_threats("5_1_1")
} # }